Add BaseConcatModel to remove all the redundant methods
